python-backend/requirements.txt
flask
flask-cors
faiss-cpu
sentence-transformers
numpy
mss
pytesseract
Pillow
requests
